requirement.txt
absl-py==0.10.0
asgiref==3.2.10
astunparse==1.6.3
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
enum34==1.1.10
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
opt-einsum==3.3.0
pandas==1.1.1
Pillow==7.2.0
plaidbench==0.7.0
plaidml==0.7.0
plaidml-keras==0.7.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
